Refactor and handle loading
