--- frontend/src/components/AccessProjectDashboardButton.tsx
@@ -1,55 +1,0 @@
-'use client'
-
-import { useQuery } from '@apollo/client'
-import { Button } from '@heroui/button'
-import Link from 'next/link'
-import { useSession } from 'next-auth/react'
-import { FC, useState, useEffect } from 'react'
-import { handleAppError } from 'app/global-error'
-import { GET_USER_IS_OWASP_STAFF } from 'server/queries/userQueries'
-import type { User } from 'types/user'
-import { userAuthStatus } from 'utils/constants'
-import LoadingSpinner from 'components/LoadingSpinner'
-
-const AccessProjectDashboardButton: FC = () => {
-  const [user, setUser] = useState<User>()
-  const { data: session, status } = useSession()
-  const {
-    data,
-    error: graphQLError,
-    loading,
-  } = useQuery(GET_USER_IS_OWASP_STAFF, {
-    variables: {
-      login: session?.user?.login,
-    },
-    skip: !session?.user?.login,
-  })
-
-  useEffect(() => {
-    if (graphQLError) {
-      handleAppError(graphQLError)
-    }
-    if (data) {
-      setUser(data.user)
-    }
-  }, [data, graphQLError])
-
-  if (loading || status === userAuthStatus.LOADING) {
-    return <LoadingSpinner />
-  }
-  if (user && !user.isOwaspStaff) {
-    return <></>
-  }
-
-  if (!user) {
-    return <></>
-  }
-
-  return (
-    <Link href="/projects/dashboard">
-      <Button color="secondary">Go To Dashboard</Button>
-    </Link>
-  )
-}
-
-export default AccessProjectDashboardButton
