+++ frontend/src/components/DashboardWrapper.tsx
@@ -8,15 +8,13 @@ import { handleAppError } from 'app/global-error'
 import { GET_USER_IS_OWASP_STAFF } from 'server/queries/userQueries'
 import type { User } from 'types/user'
 import { userAuthStatus } from 'utils/constants'
+import LoadingSpinner from 'components/LoadingSpinner'
 
 const DashboardWrapper: FC<{ children: ReactNode }> = ({ children }) => {
   const { status, data: session } = useSession()
   const [user, setUser] = useState<User>()
 
   useEffect(() => {
-    if (status === userAuthStatus.LOADING) {
-      return
-    }
     if (status === userAuthStatus.UNAUTHENTICATED) {
       notFound()
     }
@@ -42,10 +40,14 @@ const DashboardWrapper: FC<{ children: ReactNode }> = ({ children }) => {
   }, [data, graphQLError])
 
   if (loading || status === userAuthStatus.LOADING) {
-    return
+    return <LoadingSpinner />
   }
 
-  if (user && !user.isOwaspStaff) {
+  if (!user) {
+    return <LoadingSpinner />
+  }
+
+  if (!user.isOwaspStaff) {
     notFound()
   }
   return <>{children}</>
